Completed dsiplay class

+++ src/simulation/simulation.cpp
@@ -7,7 +7,7 @@ namespace smooth_sim
 ////////////////////////////////////////////////////////
 /////          SIMULATION CONSTRUCTOR             //////
 ////////////////////////////////////////////////////////
-Simulation::Simulation(int S, int E, int I, int R, int number_of_clusters, int number_of_Locations, double side,
+Simulation::Simulation(int S, int E, int I, int R, int number_of_clusters, int number_of_Locations, unsigned side,
                        double alpha, double gamma, double beta, double spread_radius, int time_in_minutes)
     : sim_engine{},
       wrld{side, number_of_clusters, number_of_Locations, S, E, I, R},
